+++ index.js
@@ -17,7 +17,7 @@ for (const file of commandFiles) {
 
 client.on('message', msg => {
   
-	if (!msg.content.startsWith(prefix) || msg.author.bot) return;
+	//if (!msg.content.startsWith(prefix) || msg.author.bot) return;
 
 	const args = msg.content.slice(prefix.length).trim().split(/ +/);
 	const command = args.shift().toLowerCase();
@@ -36,7 +36,15 @@ client.on('message', msg => {
 //Logs to the console when it is ready
 client.once('ready', () => {
     console.log(client.user.tag + " is live!");
-    client.user.setActivity("Chase Squirrely");
+    client.user.setActivity(`Mrs. Beasley Beezin'`, {type: "WATCHING"});
+    
+    // Log db info to console for debugging
+    /*
+    const { db } = require("./alpha-variables.js");
+    db.get("819017134105559070").then(values => {
+        console.log(values)
+    });
+    */
 });
 
 // Logs the bot into the Discord client